Fix type issue.

+++ src/libmv/image/convolve.h
@@ -34,7 +34,7 @@ inline double Gaussian(double x, double sigma) {
 }
 // 2D gaussian (zero mean)
 // (9) in http://mathworld.wolfram.com/GaussianFunction.html
-inline float Gaussian2D(double x, double y, float sigma)
+inline double Gaussian2D(double x, double y, double sigma)
 {
   return 1.0/(2.0*M_PI*sigma*sigma) * exp( -(x*x+y*y)/(2.0*sigma*sigma));
 }
